requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
Pillow
tabula-py
pandas
python-dotenv
faiss-cpu
PyMuPDF
